GenericFrontend/requirements.txt
asgiref==3.2.7
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
#boto3==1.12.36
boto3
botocore==1.15.36
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.5
django-cors-headers==3.2.1
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
facebook-sdk==3.1.0
flake8==3.7.9
google-auth==1.13.1
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jmespath==0.9.5
# jsonschema==2.6.0
jsonschema

lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.2.1
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
sendgrid==6.0.5
sentry-sdk==0.14.2
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.1.0
django_extensions
# pycryptodome
django-migrate-sql-deux
wiki
html-testRunner
--pre wiki